requirements.txt
numpy
scikit-learn
Pillow
matplotlib
